docker setup for serving static files behind CAS
The problem: part of this site should be behind CAS auth, but it's a
static generated site, so there's nothing there to do the auth.

To solve that, I built https://github.com/ccnmtl/apache2-cas-docker
which is apache2 with mod_auth_cas packaged into a docker image.

This can now build on that image with its own content.

Here's roughly how the deploy setup will work:

* jenkins does the usual step of using hugo to generate the
  site (populating `public/`) and mucking a bit with the json data.
* then it runs a 'docker build .' to "bake" that site into a docker
  image, which it pushes up to the docker hub
* staging server runs a `docker pull` to get the newer version of the
  image.
* staging server has an upstart config that basically just starts
  `docker run -e SERVER_NAME=https://gainesfilm.stage.ccnmtl.columbia.edu/ ccnmtl/gainesfilm`
* jenkins does a restart on that service after the docker pull
* same thing happens on the production server (docker pull + restart the
  service), but it's just configured with a different `SERVER_NAME` env
  variable.

the `docker-run.sh` script just replaces the `SERVER_NAME` in the apache
config at runtime so we don't have to have duplicate config files for
staging/prod and we can still bake and then deploy one single image that
will go through the pipeline to each environment. If we find that we do
the same thing on other projects that build off `ccnmtl/apache2-cas`, I
may move that technique up to the base image.